Code fixes + small upgrade
Fixed code so that the robot travels backwards when on black and forwards when camera sees complete white. Added code to make somewhat sharper turns in the maze. Coded to allow an 'error' for when the robot comes across a white line, so that it continues forward.  Attempted to use boolean variables to split code into quadrants. 

AV5 the current working version. Was some confusion and fixes which resulted in v 2,3,4. Very messy versioning control so far. 

Robot still getting stuck on one corner in quadrant 3 and has yet to be programmed to open the gate in quadrant 1.